igreja/requirements.txt
ttkbootstrap
yt-dlp
python-docx
faster-whisper
pillow
rawpy
tkinterdnd2
pypdf
PyMuPDF
requests
beautifulsoup4
